pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-channel-multiplexer"
version = "0.1.0"
description = "Telegram bot that forwards messages from source channels to multiple targets."
authors = [
  { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "aiogram>=3.3.0,<4.0.0",
  "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.21",
  "pytest-mock>=3.11"
]

[project.scripts]
telegram-channel-multiplexer = "telegram_channel_multiplexer.bot:run"

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "strict"
